Safe-guard when destroying pipelines

+++ oblo/renderer/src/oblo/renderer/draw/pass_manager.cpp
@@ -354,22 +354,18 @@ namespace oblo
 
     namespace
     {
-        void destroy_pipeline(gpu::gpu_instance& ctx, const render_pipeline& variant)
+        template <typename T>
+        void destroy_pipeline(gpu::gpu_instance& ctx, const T& variant)
         {
-            ctx.destroy_deferred(variant.pipeline, ctx.get_submit_index());
-            ctx.destroy_deferred(variant.descriptorSetLayout, ctx.get_submit_index());
-        }
+            if (variant.pipeline)
+            {
+                ctx.destroy_deferred(variant.pipeline, ctx.get_submit_index());
+            }
 
-        void destroy_pipeline(gpu::gpu_instance& ctx, const compute_pipeline& variant)
-        {
-            ctx.destroy_deferred(variant.pipeline, ctx.get_submit_index());
-            ctx.destroy_deferred(variant.descriptorSetLayout, ctx.get_submit_index());
-        }
-
-        void destroy_pipeline(gpu::gpu_instance& ctx, const raytracing_pipeline& variant)
-        {
-            ctx.destroy_deferred(variant.pipeline, ctx.get_submit_index());
-            ctx.destroy_deferred(variant.descriptorSetLayout, ctx.get_submit_index());
+            if (variant.descriptorSetLayout)
+            {
+                ctx.destroy_deferred(variant.descriptorSetLayout, ctx.get_submit_index());
+            }
         }
 
         template <typename Pass, typename Pipelines>
